fix: Правки по ревью 1.

+++ src/main/java/ru/yandex/practicum/filmorate/model/Mpa.java
@@ -7,6 +7,6 @@ import lombok.Data;
 @Builder
 public class Mpa {
 
-    long id;
+    Integer id;
     String name;
 }
